requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
openai>=0.27.0  # For OpenAI GPT
anthropic>=0.5.0  # For Anthropic Claude
google-generativeai>=0.3.0  # For Google Gemini
langchain>=0.1.0  # For multi-LLM orchestration
scipy>=1.7.0  # For statistical tests